+++ src/components/nav/nav-text.spec.js
@@ -5,20 +5,7 @@ describe('nav > nav-text', () => {
   it('has expected default structure', async () => {
     const wrapper = mount(BNavText)
 
-    expect(wrapper.is('span')).toBe(true)
-    expect(wrapper.classes()).toContain('navbar-text')
-    expect(wrapper.classes().length).toBe(1)
-    expect(wrapper.text()).toEqual('')
-  })
-
-  it('renders custom root element when prop tag is set', async () => {
-    const wrapper = mount(BNavText, {
-      propsData: {
-        tag: 'div'
-      }
-    })
-
-    expect(wrapper.is('div')).toBe(true)
+    expect(wrapper.is('li')).toBe(true)
     expect(wrapper.classes()).toContain('navbar-text')
     expect(wrapper.classes().length).toBe(1)
     expect(wrapper.text()).toEqual('')
@@ -31,7 +18,7 @@ describe('nav > nav-text', () => {
       }
     })
 
-    expect(wrapper.is('span')).toBe(true)
+    expect(wrapper.is('li')).toBe(true)
     expect(wrapper.classes()).toContain('navbar-text')
     expect(wrapper.classes().length).toBe(1)
     expect(wrapper.text()).toEqual('foobar')