+++ src/components/table/tbody.js
@@ -53,6 +53,12 @@ export const BTbody = /*#__PURE__*/ Vue.extend({
       // Sticky headers are only supported in thead
       return false
     },
+    hasStickyHeader() {
+      // Sniffed by <b-tr> / <b-td> / <b-th>
+      // Needed to handle header background classes, due to lack of
+      // background color inheritance with Bootstrap v4 table CSS
+      return !this.isStacked && this.bvTable.stickyHeader
+    },
     tableVariant() /* istanbul ignore next: Not currently sniffed in tests */ {
       // Sniffed by <b-tr> / <b-td> / <b-th>
       return this.bvTable.tableVariant
@@ -65,21 +71,25 @@ export const BTbody = /*#__PURE__*/ Vue.extend({
     },
     tbodyProps() {
       return this.tbodyTransitionProps ? { ...this.tbodyTransitionProps, tag: 'tbody' } : {}
-    },
-    tbodyListeners() {
-      const handlers = this.tbodyTransitionHandlers || {}
-      return { ...this.$listeners, ...handlers }
     }
   },
   render(h) {
+    const data = {
+      props: this.tbodyProps,
+      attrs: this.tbodyAttrs
+    }
+    if (this.isTransitionGroup) {
+      // We use native listeners if a transition group
+      // for any delegated events
+      data.on = this.tbodyTransitionHandlers || {}
+      data.nativeOn = this.$listeners || {}
+    } else {
+      // Otherwise we place any listeners on the tbody element
+      data.on = this.$listeners || {}
+    }
     return h(
       this.isTransitionGroup ? 'transition-group' : 'tbody',
-      {
-        props: this.tbodyProps,
-        attrs: this.tbodyAttrs,
-        // Pass down any listeners
-        on: this.tbodyListeners
-      },
+      data,
       this.normalizeSlot('default', {})
     )
   }