+++ src/mixins/focus-in.js
@@ -1,5 +1,7 @@
 import { eventOff, eventOn } from '../utils/dom'
 
+const eventOptions = { passive: true, capture: false }
+
 // @vue/component
 export default {
   data() {
@@ -10,9 +12,9 @@ export default {
   watch: {
     listenForFocusIn(newValue, oldValue) {
       if (newValue !== oldValue) {
-        eventOff(this.focusInElement, 'focusin', this._focusInHandler, false)
+        eventOff(this.focusInElement, 'focusin', this._focusInHandler, eventOptions)
         if (newValue) {
-          eventOn(this.focusInElement, 'focusin', this._focusInHandler, false)
+          eventOn(this.focusInElement, 'focusin', this._focusInHandler, eventOptions)
         }
       }
     }
@@ -26,11 +28,11 @@ export default {
       this.focusInElement = document
     }
     if (this.listenForFocusIn) {
-      eventOn(this.focusInElement, 'focusin', this._focusInHandler, false)
+      eventOn(this.focusInElement, 'focusin', this._focusInHandler, eventOptions)
     }
   },
   beforeDestroy() /* istanbul ignore next */ {
-    eventOff(this.focusInElement, 'focusin', this._focusInHandler, false)
+    eventOff(this.focusInElement, 'focusin', this._focusInHandler, eventOptions)
   },
   methods: {
     _focusInHandler(evt) {